hw16

+++ students/views.py
@@ -1,25 +1,12 @@
 # -*- coding: utf-8 -*-
-from django.shortcuts import render, redirect
+from django.shortcuts import render, redirect, get_object_or_404
 from students.models import Student, Group
-from courses.models import Course
 from django import forms
 
 
-class StudentForm(forms.Form):
-    PACKAGE_CHOISES = (
-        ('standart', 'Standart'),
-        ('gold', 'Glod'),
-        ('platimun', 'Platinum'),
-    )
-    name = forms.CharField(max_length=225)
-    surname = forms.CharField(max_length=225)
-    date_of_birth = forms.DateField(required=False, label='birthday')
-    email = forms.EmailField()
-    phone = forms.CharField(max_length=15, required=False)
-    courses = forms.ModelMultipleChoiceField(
-    										 queryset=Course.objects.all())
-    package = forms.ChoiceField(choices=PACKAGE_CHOISES)
-    group = forms.ModelChoiceField(queryset=Group.objects.all(), required=False)
+class StudentForm(forms.ModelForm):
+    class Meta:
+    	model = Student
 
 
 def students_list(request):
@@ -31,58 +18,36 @@ def student_item(request, student_id):
     return render(request, 'students/item.html', {'student': student},)
 
 def student_edit(request, student_id):
+	title = "Student edit"
 	student = Student.objects.get(id=student_id)
 	if request.method == 'POST':
-		form = StudentForm(request.POST)
+		form = StudentForm(request.POST, instance=student)
 		if form.is_valid():
-			student.name = form.cleaned_data('name')
-			student.surname = form.cleaned_data('surname')
-			student.date_of_birth = form.cleaned_data('date_of_birth')
-			student.email = form.cleaned_data('email')
-			student.phone = form.cleaned_data('phone')
-			student.courses = form.cleaned_data('courses')
-			student.package = form.cleaned_data('package')
-			student.group = form.cleaned_data('group')
-			student.save()
-			# здесь, не понимаю, в чем ошибка
-			return redirect('student_edit', student.id,)
+			student = form.save()
+			return redirect('student_item', student.id,)
 	else:
-		form = StudentForm(initial={
-		   'name': student.name,
-	       'surname': student.surname,
-	       'date_of_birth': student.date_of_birth,
-	       'email': student.email,
-	       'phone': student.phone,
-	       # 'courses': student.courses,
-	       # c этим полем не выходит... как правильно, не пойму
-	       'package': student.package,
-	       'group': student.group,
-	})
+		form = StudentForm(instance=student)
 	return render(request, 'students/edit.html',
 		                   {'form': form,
-		                   'student': student})
+		                   'student': student,
+		                   'title': title})
 
 def student_new(request):
 	title = 'Add new student'
 	if request.method == "POST":
 		form = StudentForm(request.POST)
 		if form.is_valid():
-			student.name = form.cleaned_data('name')
-			student.surname = form.cleaned_data('surname')
-			student.date_of_birth = form.cleaned_data('date_of_birth')
-			student.email = form.cleaned_data('email')
-			student.phone = form.cleaned_data('phone')
-			student.courses = form.cleaned_data('courses')
-			student.package = form.cleaned_data('package')
-			student.group = form.cleaned_data('group')
-			student.save()			
+			form.save()			
 			return redirect('students_list')
 	else:
 		form = StudentForm()
 	return render(request, 'students/edit.html', {'form': form,
-												 'title': title})
+												 'title': title,
+												 })
 
 def student_delete(request, student_id):
-	student = Student.objects.get(id=student_id)
-	student.delete()
+	student = get_object_or_404(Student, pk=student_id)
+	#student = Student.objects.get(id=student_id)
+	if request.method == "GET":
+		student.delete()
 	return redirect('students_list')
